reorganize: move mission to ai directory
