+++ src/dct/ai/Commander.lua
@@ -9,7 +9,7 @@ local utils      = require("libs.utils")
 local containers = require("libs.containers")
 local enum       = require("dct.enum")
 local dctutils   = require("dct.utils")
-local Mission    = require("dct.Mission")
+local Mission    = require("dct.ai.Mission")
 local Stats      = require("dct.Stats")
 local Command    = require("dct.Command")
 local Logger     = require("dct.Logger").getByName("Commander")
